chore(backend): move encryption util methods inside core folder, rename DTO folder to DTOs

+++ server-backend/NutritionTracker.Api/DTOs/FoodNutritionDto.cs
@@ -5,9 +5,9 @@
         public string FoodName { get; set; }
         public double QuantityInGrams { get; set; }
 
-        public double Calories { get; set; }
-        public double Protein { get; set; }
-        public double Carbs { get; set; }
-        public double Fat { get; set; }
+        public float Calories { get; set; }
+        public float Protein { get; set; }
+        public float Carbs { get; set; }
+        public float Fat { get; set; }
     }
 }
